+++ trionyx/settings.py
@@ -91,6 +91,7 @@ MIDDLEWARE = (
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
     'trionyx.trionyx.middleware.LoginRequiredMiddleware',
+    'trionyx.trionyx.middleware.GlobalRequestMiddleware',
 )
 
 # ==============================================================================